feat(ValueBox): 增加浮点数相等性语义支持，优化 Update 方法以处理 -0.0 和 NaN 的比较

+++ benchmarks/StateJournal4.Benchmarks/SJValueLayoutBenchmarks.cs
@@ -120,7 +120,7 @@ public class SJValueLayoutBenchmarks {
             int roll = rng.Next(100);
             if (roll < 50) {
                 // Inline-Nonneg-Integer: 高位 01, 低 62 位是值
-                bits = 0x4000_0000_0000_0000UL | (ulong)rng.Next(0, 10000);
+                bits = 0x4000_0000_0000_0000UL | (ulong)(uint)rng.Next(0, 10000);
             } else if (roll < 80) {
                 // Lossy-Double: 高位 1
                 bits = unchecked((ulong)rng.NextInt64()) | 0x8000_0000_0000_0000UL;
